+++ backend/src/auth/auth.middleware.rs
@@ -1,0 +1,37 @@
+use actix_web::dev::ServiceRequest;
+use actix_web_httpauth::extractors::{bearer::{self, BearerAuth}, AuthenticationError};
+use hmac::{Hmac, Mac};
+use sha2::Sha256;
+use jwt::VerifyWithKey;
+use actix_web::{HttpMessage, Error};
+
+use super::auth_models::TokenClaims;
+
+pub async fn token_validator(
+    req: ServiceRequest,
+    credentials: BearerAuth,
+) -> Result<ServiceRequest, (Error, ServiceRequest)> {
+    let jwt_secret = std::env::var("JWT_SECRET").expect("JWT secret not found");
+    let key: Hmac<Sha256> = Hmac::new_from_slice(jwt_secret.as_bytes()).unwrap();
+    let token_string = credentials.token();
+
+    let claims: Result<TokenClaims, &str> = token_string
+        .verify_with_key(&key)
+        .map_err(|_| "Invalid token");
+
+    match claims {
+        Ok(value) => {
+            req.extensions_mut().insert(value);
+            Ok(req)
+        }
+        Err(_) => {
+            let config = req
+                .app_data::<bearer::Config>()
+                .cloned()
+                .unwrap_or_default()
+                .scope("");
+
+            Err((AuthenticationError::from(config).into(), req))
+        }
+    }
+}
