+++ backend/src/auth/mod.rs
@@ -1,8 +1,8 @@
 #[path = "auth.control.rs"]
 pub mod auth_controllers;
 
-#[path = "auth.service.rs"]
-pub mod auth_services;
+#[path = "auth.middleware.rs"]
+pub mod auth_middleware;
 
 #[path = "auth.models.rs"]
 pub mod auth_models;
