--- backend/src/services/mod.rs
@@ -1,2 +1,0 @@
-#[path = "auth.service.rs"]
-pub mod auth;
